+++ examples/character_controller.rs
@@ -1,5 +1,9 @@
 //! Demonstrates how to create a simple platforming 2D character controller
 //! using actions with both keyboard and gamepad controls.
+//!
+//! For kinematic character controllers, input should be accumulated and applied
+//! to physics in a fixed timestep as recommended in [this Bevy example](https://bevy.org/examples/movement/physics-in-fixed-timestep/)
+//! and as used in [bevy_ahoy](https://github.com/janhohenheim/bevy_ahoy).
 
 use bevy::prelude::*;
 use bevy_enhanced_input::prelude::*;
@@ -14,10 +18,25 @@ fn main() {
     App::new()
         .add_plugins((DefaultPlugins, EnhancedInputPlugin))
         .add_input_context::<Player>()
+        .init_resource::<FixedUpdateRan>()
         .add_systems(Startup, setup)
-        .add_systems(Update, calculate_physics)
-        .add_observer(apply_movement)
-        .add_observer(apply_jump)
+        .add_systems(PreUpdate, reset_fixed_update_ran)
+        .add_systems(FixedPreUpdate, set_fixed_update_ran)
+        // Apply input before advancing physics
+        .add_systems(FixedUpdate, apply_input)
+        .add_systems(FixedPostUpdate, advance_physics)
+        .add_systems(
+            // Run outside the schedule loop that repeats FixedMain zero-to-many times
+            RunFixedMainLoop,
+            clear_input
+                // To prevent prematurely clearing input, only clear if
+                // FixedUpdate ran one-to-many times during this loop
+                .run_if(fixed_update_ran)
+                // Run *after* loop
+                .in_set(RunFixedMainLoopSystems::AfterFixedMainLoop),
+        )
+        .add_observer(accumulate_movement)
+        .add_observer(accumulate_jump)
         .run();
 }
 
@@ -41,10 +60,12 @@ fn setup(
         MeshMaterial2d(materials.add(Color::srgb(1.0, 0.0, 0.5))),
         Transform::from_translation(Vec3::Y * (GROUND_LEVEL + 500.0)),
         PlayerPhysics::default(),
+        AccumulatedInput::default(),
         actions!(Player[
             (
                 Action::<Movement>::new(),
                 DeadZone::default(),
+                // Velocity is set to a smoothed input to simulate acceleration.
                 SmoothNudge::default(),
                 // We don't apply `DeltaScale` here because the movement vector is
                 // multiplied by delta time during the physics calculation.
@@ -63,25 +84,40 @@ fn setup(
     ));
 }
 
-fn apply_movement(movement: On<Fire<Movement>>, mut query: Query<&mut PlayerPhysics>) {
-    let mut physics = query.get_mut(movement.context).unwrap();
-    physics.velocity.x = movement.value;
+fn accumulate_movement(movement: On<Fire<Movement>>, mut inputs: Query<&mut AccumulatedInput>) {
+    let mut accumulated_inputs = inputs.get_mut(movement.context).unwrap();
+    accumulated_inputs.movement = movement.value;
 }
 
-fn apply_jump(jump: On<Fire<Jump>>, mut query: Query<&mut PlayerPhysics>) {
-    let mut physics = query.get_mut(jump.context).unwrap();
-    if physics.is_grounded {
-        // Jump only if on the ground.
-        physics.velocity.y = JUMP_VELOCITY;
-        physics.is_grounded = false;
+fn accumulate_jump(jump: On<Fire<Jump>>, mut inputs: Query<&mut AccumulatedInput>) {
+    let mut accumulated_inputs = inputs.get_mut(jump.context).unwrap();
+    accumulated_inputs.jump = true;
+}
+
+fn clear_input(mut inputs: Query<&mut AccumulatedInput>) {
+    for mut inputs in &mut inputs {
+        *inputs = Default::default();
     }
 }
 
-fn calculate_physics(time: Res<Time>, mut query: Query<(&mut Transform, &mut PlayerPhysics)>) {
-    for (mut transform, mut physics) in query.iter_mut() {
-        physics.velocity.y -= GRAVITY * time.delta_secs();
-        transform.translation.y += physics.velocity.y * time.delta_secs();
-        transform.translation.x += physics.velocity.x * time.delta_secs();
+fn apply_input(players: Query<(&mut PlayerPhysics, &AccumulatedInput)>) {
+    for (mut physics, input) in players {
+        physics.velocity.x = input.movement;
+        if input.jump && physics.is_grounded {
+            physics.velocity.y = JUMP_VELOCITY;
+            physics.is_grounded = false;
+        }
+    }
+}
+
+fn advance_physics(
+    fixed_time: Res<Time<Fixed>>,
+    mut players: Query<(&mut Transform, &mut PlayerPhysics)>,
+) {
+    for (mut transform, mut physics) in &mut players {
+        physics.velocity.y -= GRAVITY * fixed_time.delta_secs();
+        transform.translation.y += physics.velocity.y * fixed_time.delta_secs();
+        transform.translation.x += physics.velocity.x * fixed_time.delta_secs();
 
         // Prevent moving off screen.
         const MAX_X: f32 = GROUND_WIDTH / 2.0 - PLAYER.x / 2.0;
@@ -113,3 +149,26 @@ struct Movement;
 #[derive(Debug, InputAction)]
 #[action_output(bool)]
 struct Jump;
+
+/// Accumulated input since the last fixed update.
+#[derive(Component, Default)]
+struct AccumulatedInput {
+    movement: f32,
+    jump: bool,
+}
+
+/// True if FixedPreUpdate was run this frame.
+#[derive(Resource, Deref, DerefMut, Default)]
+struct FixedUpdateRan(bool);
+
+fn reset_fixed_update_ran(mut ran: ResMut<FixedUpdateRan>) {
+    **ran = false;
+}
+
+fn set_fixed_update_ran(mut ran: ResMut<FixedUpdateRan>) {
+    **ran = true;
+}
+
+fn fixed_update_ran(ran: Res<FixedUpdateRan>) -> bool {
+    **ran
+}