+++ src/binding.rs
@@ -105,6 +105,11 @@ pub enum Binding {
     /// inputs, not just the one that activated it. To have an action with this binding
     /// evaluated first, place it in a higher-priority context.
     AnyKey,
+    /// Identifier of the input entry in [`CustomInputs`] to read.
+    ///
+    /// Use this for inputs that can't be represented or feed into Bevy input resources.
+    /// Obtain an identifier with [`CustomInputs::register_input`].
+    Custom(CustomInput),
     /// Doesn't correspond to any input, captured as [`ActionValue::Bool`] with `false`.
     ///
     /// Useful for expressing empty bindings in [presets](crate::preset).
@@ -145,6 +150,7 @@ impl Binding {
             Binding::GamepadButton(_)
             | Binding::GamepadAxis(_)
             | Binding::AnyKey
+            | Binding::Custom(_)
             | Binding::None => ModKeys::empty(),
         }
     }
@@ -175,6 +181,7 @@ impl Display for Binding {
             Binding::GamepadButton(gamepad_button) => write!(f, "{gamepad_button:?}"),
             Binding::GamepadAxis(gamepad_axis) => write!(f, "{gamepad_axis:?}"),
             Binding::AnyKey => write!(f, "Any Key"),
+            Binding::Custom(input) => write!(f, "{input:?}"),
             Binding::None => write!(f, "None"),
         }
     }
@@ -232,7 +239,8 @@ impl<I: Into<Binding>> InputModKeys for I {
             Binding::GamepadButton { .. }
             | Binding::GamepadAxis { .. }
             | Binding::None
-            | Binding::AnyKey => {
+            | Binding::AnyKey
+            | Binding::Custom(_) => {
                 error!("can't add `{mod_keys:?}` to `{binding:?}`");
                 binding
             }